Add tests for instead of question

+++ client/src/main/java/client/scenes/questioncontrollers/OpenQuestionCtrl.java
@@ -17,7 +17,7 @@ public class OpenQuestionCtrl implements QuestionRequirements {
     private MainCtrl mainCtrl;
     private QuestionFrameCtrl questionFrameCtrl;
     private Question question;
-    private int answer;
+    private long answer;
 
     @FXML
     Button submitButton;
@@ -55,8 +55,8 @@ public class OpenQuestionCtrl implements QuestionRequirements {
     public void submit() {
         String ans = entryField.getText();
         Scanner scanner = new Scanner(ans);
-        if (scanner.hasNextInt()) {
-            this.answer = scanner.nextInt();
+        if (scanner.hasNextLong()) {
+            this.answer = scanner.nextLong();
             errorMessage.setVisible(false);
             submitButton.setText("Submitted!");
             submitButton.setDisable(true);
@@ -115,26 +115,56 @@ public class OpenQuestionCtrl implements QuestionRequirements {
 
     ;
 
+    /**
+     * Returns the main controller of the screen
+     *
+     * @return the main front-end controller
+     */
     public MainCtrl getMainCtrl() {
         return mainCtrl;
     }
 
+    /**
+     * Returns the question of the screen
+     *
+     * @return the Question that is shown in the screen
+     */
     public Question getQuestion() {
         return question;
     }
 
-    public int getAnswer() {
+    /**
+     * Returns the answer to the question
+     *
+     * @return the correct consumption value as a long
+     */
+    public long getAnswer() {
         return answer;
     }
 
+    /**
+     * Returns the error message text
+     *
+     * @return the error message that is displayed when wrong input is entered
+     */
     public Text getErrorMessage() {
         return errorMessage;
     }
 
-    public void setAnswer(int answer) {
+    /**
+     * Sets the answer
+     *
+     * @param answer the correct consumption value
+     */
+    public void setAnswer(long answer) {
         this.answer = answer;
     }
 
+    /**
+     * Sets the question to be displayed in the screen
+     *
+     * @param question the Question that is shown in the screen
+     */
     public void setQuestion(Question question) {
         this.question = question;
     }
